+++ Back-end/BE/src/main/java/com/example/BE/config/CorsConfig.java
@@ -1,0 +1,16 @@
+package com.example.BE.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class CorsConfig implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("http://localhost:3001") // Thay bằng URL của ứng dụng React
+                .allowedMethods("*")
+                .allowCredentials(true);
+    }
+}
