+++ Back-end/BE/src/main/java/com/example/BE/config/OpenApiConfig.java
@@ -1,0 +1,19 @@
+package com.example.BE.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenApiConfig {
+
+    private String appVersion = "v1.1.1";
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI().info(new Info().title("API Documentation").version(appVersion)
+                .description("API documentation for my Spring Boot project"));
+    }
+
+}
