--- Back-end/BE/src/main/java/com/example/BE/entity/CompositeClassUser.java
@@ -1,18 +1,0 @@
-package com.example.BE.entity;
-
-import jakarta.persistence.Embeddable;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-@Embeddable
-public class CompositeClassUser {
-	
-	@ManyToOne
-    @JoinColumn(name = "userid")
-    private User user;
-
-    
-	@ManyToOne
-    @JoinColumn(name = "classid")
-    private Class class_object;
-}
