+++ Back-end/BE/src/main/java/com/example/BE/exception/NotFoundException.java
@@ -1,0 +1,11 @@
+package com.example.BE.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
