+++ Back-end/BE/src/main/java/com/example/BE/model/dto/StatusEnum.java
@@ -1,0 +1,6 @@
+package com.example.BE.model.dto;
+
+public enum StatusEnum {
+    SUCCESS, ERROR
+}
+
