+++ Back-end/BE/src/main/java/com/example/BE/model/entity/CompositeClassUser.java
@@ -1,0 +1,18 @@
+package com.example.BE.model.entity;
+
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+
+@Embeddable
+public class CompositeClassUser {
+	
+	@ManyToOne
+    @JoinColumn(name = "userid")
+    private User user;
+
+    
+	@ManyToOne
+    @JoinColumn(name = "classid")
+    private Class class_object;
+}
